[Refactor] #34 Update file structure and rename
[Refactor] #34 Update file structure and rename
