+++ src/components/atoms/DetailTitle.tsx
@@ -2,7 +2,7 @@ import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
 import Button from '@mui/material/Button';
 
-export default function DoTitle() {
+export default function DetailTitle() {
   return (
     <Box
       sx={{
