+++ src/components/atoms/index.ts
@@ -1,0 +1,11 @@
+export { default as CustomButton } from './CustomButton';
+export { default as Days } from './Days';
+export { default as DetailTitle } from './DetailTitle';
+export { default as LoginBody } from './LoginBody';
+export { default as LogoutButton } from './LogoutButton';
+export { default as ModalAtom } from './ModalAtom';
+export { default as MyInfo } from './MyInfo';
+export { default as MypageLayout } from './MypageLayout';
+export { default as RoutineDate } from './RoutineDate';
+export { default as RoutinePercent } from './RoutinePercent';
+export { default as TextFiled } from './TextField';
